% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_cov_model.R
\name{brms_cov_model}
\alias{brms_cov_model}
\title{Run a brms model containing a covariance matrix}
\usage{
brms_cov_model(
  Data,
  Response,
  FixedEffect,
  ID,
  Matrix,
  Chainset = 1,
  Family = "gaussian",
  Seed = NULL,
  Trials = NA,
  PriorSamples = TRUE
)
}
\arguments{
\item{Data}{A data frame containing the data - covariates should be centered to a the mean or to a meaningful zero (see Schielzeth H. 2010. Simple means to improve the interpretability of regression coefficients. Methods Ecol Evol. 1:103–113. doi:10.1111/j.2041-210X.2010.00012.x.).}

\item{Response}{String with the name of the column in Data containing the response variable (e.g. "mass").}

\item{FixedEffect}{String with the name of the column in Data containing the fixed effect variable (e.g. "height"). To add multiple fixed effects, use c() (e.g. c("height", "sex")).}

\item{ID}{String with the name of the column in Data containing the ID of correlated individuals/groups. The current package version allows the use of a single random effect, so no additional random effects are possible yet.}

\item{Matrix}{A matrix containing the IDs as row/column names and covariance values (ex. A pedigree relatedness matrix).}

\item{Chainset}{Defines the number of iterations. Start with Chainset = 1 and increase as needed until convergence. The value of Chainsetis multiplied by 15000 in warmup, 30000 in iterations and 15 in thin intervale. For quick tests use Chainset = 0 (warmup=10; iter=110; thin=10; chains=2)}

\item{Family}{String to define the family function in the brms model. Current supported families: "gaussian", "binomial", "poisson".}

\item{Seed}{Numeric and optional. Set a seed in order to repeat the results from the model when running it more than once.}

\item{Trials}{The total number of trials in a binomial model. The number of successes should be imputed on Response.}

\item{PriorSamples}{Logical value that defines if brmsfit will contain the priors used. Default is set to \code{FALSE}, which does not includ the priors in the brmsfit.}
}
\value{
Returns a brmsfit
}
\description{
Run a brms model containing a covariance matrix
}
