% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_model.R
\name{brms_model}
\alias{brms_model}
\title{Run a brms model}
\usage{
brms_model(
  Data,
  Response,
  FixedEffect,
  RandomEffect = NULL,
  RandomSlope = NULL,
  Chainset = 1,
  Family = "gaussian",
  Seed = NULL,
  Trials = NULL,
  PriorSamples = TRUE
)
}
\arguments{
\item{Data}{A data frame containing the data - covariates should be centered to a the mean or to a meaningful zero (see Schielzeth H. 2010. Simple means to improve the interpretability of regression coefficients. Methods Ecol Evol. 1:103–113. doi:10.1111/j.2041-210X.2010.00012.x.).}

\item{Response}{String with the name of the column in Data containing the response variable (e.g. "mass").}

\item{FixedEffect}{String with the name of the column in Data containing the fixed effect variable (e.g. "height"). To add multiple fixed effects, use c() (e.g. c("height", "sex")).}

\item{RandomEffect}{String with the name of the column in Data containing the random effect variable (e.g. "species"). The current package version allows the use of a single random effect.}

\item{RandomSlope}{Vector containing two values: first a string with the name of the column in Data containing the covariate to be added as a random slope, second the random effect to which the  random slope will be applied (e.g. RandomSlope = c("height", "species")). The current package version allows the use of a single random slope.}

\item{Chainset}{Defines the number of iterations. Start with Chainset = 1 and increase as needed until convergence. The value of Chainsetis multiplied by 15000 in warmup, 30000 in iterations and 15 in thin intervale. For quick tests use Chainset = 1 (warmup=10; iter=110; thin=10; chains=2)}

\item{Family}{String to define the family function in the brms model. Current supported families: "gaussian", "binomial", "poisson".}

\item{Seed}{Numeric and optional. Set a seed in order to repeat the results from the model when running it more than once.}

\item{Trials}{The total number of trials in a binomial model. The number of successes should be imputed on Response.}

\item{PriorSamples}{Logical value that defines if brmsfit will contain the priors used. Default is set to \code{FALSE}, which does not includ the priors in the brmsfit.}
}
\value{
Returns a brmsfit
}
\description{
Run a brms model
}
\examples{
# Simulate data
md = tibble::tibble(
  group = factor(sample(1:10, 1000, replace = TRUE)),
  f_var = factor(sample(1:3, 1000, replace = TRUE)),
  n_var = rnorm(1000, mean = 0, sd = 1),
  resp = rnorm(1000, mean = 10, sd = 3))

# Run model
mod = brms_model(Response = "resp", 
                 FixedEffect = c("f_var","n_var"), 
                 RandomEffect = "group", 
                 Family = "gaussian", 
                 Data = md)




}
