% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_slopes.R
\name{compare_slopes}
\alias{compare_slopes}
\title{Compare slope values of different brms models}
\usage{
compare_slopes(..., Slope)
}
\arguments{
\item{...}{brms models to be compared}

\item{Slope}{A string containing the name of the covariate to have the slopes compared between models}
}
\value{
Returns a table of values with the log(slope model 1/slope model 2)
}
\description{
Compare slope values of different brms models
}
\examples{

# Simulate data
md = tibble::tibble(
  group = factor(sample(1:10, 1000, replace = TRUE)),
  f_var = factor(sample(1:3, 1000, replace = TRUE)),
  n_var = rnorm(1000, mean = 0, sd = 1),
  resp = rnorm(1000, mean = 10, sd = 3))

# Run model without random slope
mod = brms_model(Response = "resp", 
                 FixedEffect = c("f_var","n_var"), 
                 RandomEffect = "group", 
                 Family = "gaussian", 
                 Data = md)

# Run model with random slope
mod_RS = brms_model(Response = "resp", 
                 FixedEffect = c("f_var","n_var"), 
                 RandomSlope = c("n_var", "group"),
                 Family = "gaussian", 
                 Data = md)

compare_slopes(mod, mod_RS, Slope = "n_var")

}
