% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R, R/mod_cov.R, R/mod_re.R, R/mod_rs.R
\docType{data}
\name{md}
\alias{md}
\title{mod
model with fixed effects only}
\format{
brmsfit

brmsfit

brmsfit

brmsfit
}
\usage{
md

md

md

md
}
\description{
mod
model with fixed effects only

mod_cov
model with fixed effects and a random effect with a covariance matrix

mod_RE
model with fixed effects and random effects

mod_RS
model with fixed effects and a random slope
}
\keyword{datasets}
