% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R, R/model-examples.R, R/model_examples.R
\docType{data}
\name{mod}
\alias{mod}
\title{Example model with fixed effects only}
\format{
A \code{brmsfit} object.

Several \code{brmsfit} objects.

A \code{brmsfit} object.
}
\source{
Generated using the \code{brms_model} function.

Generated using the \code{brms_model} and \code{brms_cov_model} functions.

Generated using the \code{brms_model} function.
}
\usage{
mod

mod

mod
}
\description{
Example model with fixed effects only

Several model examples that can be used to test other functions of \code{VarDecomp}

Model example that can be used to test other functions of \code{VarDecomp}. A model with fixed effects only
}
\keyword{datasets}
