% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cov.R, R/model_examples-mod_cov.R,
%   R/model_examples.R
\docType{data}
\name{mod_cov}
\alias{mod_cov}
\title{model with fixed effects and a random effect with a covariance matrix}
\format{
A \code{brmsfit} object.

A \code{brmsfit} object.

A \code{brmsfit} object.
}
\source{
Generated using the \code{brms_cov_model} function.

Generated using the \code{brms_model} function.

Generated using the \code{brms_model} function.
}
\usage{
mod_cov

mod_cov

mod_cov
}
\description{
model with fixed effects and a random effect with a covariance matrix

Model example that can be used to test other functions of \code{VarDecomp}. A model with a covariance matrix.

Model example that can be used to test other functions of \code{VarDecomp}. A model with a covariance matrix.
}
\keyword{datasets}
