% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_re.R
\docType{data}
\name{mod_RE}
\alias{mod_RE}
\title{Example model with fixed and random effects}
\format{
A \code{brmsfit} object.
}
\source{
Generated using the \code{brms_model} function.
}
\usage{
mod_RE
}
\description{
Example model with fixed and random effects
}
\keyword{datasets}
