% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_rs.R
\docType{data}
\name{mod_RS}
\alias{mod_RS}
\title{Example model with fixed effects and random slopes}
\format{
A \code{brmsfit} object.
}
\source{
Generated using the \code{brms_model} function.
}
\usage{
mod_RS
}
\description{
Example model with fixed effects and random slopes
}
\keyword{datasets}
