% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{Perform model fit checks for brms models}
\usage{
model_fit(brmsfit, Prior = FALSE)
}
\arguments{
\item{brmsfit}{The output of a brms model. You can use VarDecomp::brms_model() to produce a brmsfit.}

\item{Prior}{A logical argument defining whether the \code{brmsfit} contains prior samples. If set to \code{TRUE} it will produce plots comparing the log distributions of priors and posterior samples for each covariate.}
}
\value{
Returns a list containing (a) the maximum R-hat value, (b) the minimum effective sample size, (c) traceplots, (d) posterior predictive check plots, and (e) prior and posterior sample plots (if priors are available).
}
\description{
Perform model fit checks for brms models
}
\examples{

# Simulate data
md = tibble::tibble(
  group = factor(sample(1:10, 1000, replace = TRUE)),
  f_var = factor(sample(1:3, 1000, replace = TRUE)),
  n_var = rnorm(1000, mean = 0, sd = 1),
  resp = rnorm(1000, mean = 10, sd = 3))

# Run model
mod = brms_model(Response = "resp", 
                 FixedEffect = c("f_var","n_var"), 
                 RandomEffect = "group", 
                 Family = "gaussian", 
                 Data = md)


# Check model fit 
model_fit(mod, Prior = TRUE)

}
