% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intervals.R
\name{plot_intervals}
\alias{plot_intervals}
\title{Plot probability density intervals for fixed effects of brms models}
\usage{
plot_intervals(brmsfit)
}
\arguments{
\item{brmsfit}{The output of a brms model. You can use VarDecomp::brms_model() to produce a brmsfit.}
}
\value{
Returns a density plot with z-transformed fixed effect estimates.
}
\description{
Plot probability density intervals for fixed effects of brms models
}
\examples{

# Simulate data
md = tibble::tibble(
  group = factor(sample(1:10, 1000, replace = TRUE)),
  f_var = factor(sample(1:3, 1000, replace = TRUE)),
  n_var = rnorm(1000, mean = 0, sd = 1),
  resp = rnorm(1000, mean = 10, sd = 3))

# Run model
mod = brms_model(Response = "resp", 
                 FixedEffect = c("f_var","n_var"), 
                 RandomEffect = "group", 
                 Family = "gaussian", 
                 Data = md)

# Plot fixed effects
plot_intervals(mod)


}
