% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_r2.R
\name{plot_R2}
\alias{plot_R2}
\title{Plot R2 values of brms models}
\usage{
plot_R2(brmsfit, Colors = NULL, PlotType = "pizza", Label = NULL, Break = NULL)
}
\arguments{
\item{brmsfit}{The output of a brms model. You can use VarDecomp::brms_model() to produce a brmsfit.}

\item{Colors}{A vector with string of the colors to be attributed to each variable ordered by R2, with residual at the end (e.g. c("blue", "green", "gray")).}

\item{PlotType}{Choose if the plot should be a 'bar' or 'pizza' plot.}

\item{Label}{Choose how values should be displayed ("box" or "text"). Default is set to not display values.}

\item{Break}{Add a break in the bar plot axis in cases where a variable explains a large portion of the variance.}
}
\value{
Returns a bar plot with the proportion of variance explained by each variable.
}
\description{
Plot R2 values of brms models
}
\examples{

# Simulate data
md = tibble::tibble(
  group = factor(sample(1:10, 1000, replace = TRUE)),
  f_var = factor(sample(1:3, 1000, replace = TRUE)),
  n_var = rnorm(1000, mean = 0, sd = 1),
  resp = rnorm(1000, mean = 10, sd = 3))

# Run model
mod = brms_model(Response = "resp", 
                 FixedEffect = c("f_var","n_var"), 
                 RandomEffect = "group", 
                 Family = "gaussian", 
                 Data = md)

# Plot R2
plot_R2(mod, PlotType = "pizza", Label = "box")


}
