% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_decomp.R
\name{var_decomp}
\alias{var_decomp}
\title{Variance decomposition for brms models}
\usage{
var_decomp(brmsfit)
}
\arguments{
\item{brmsfit}{The output of a brms model. You can use VarDecomp::brms_model() to produce a brmsfit.}
}
\value{
Returns a data frame with posterior estimates.
}
\description{
Variance decomposition for brms models
}
\examples{

# Simulate data
md = tibble::tibble(
  group = factor(sample(1:10, 1000, replace = TRUE)),
  f_var = factor(sample(1:3, 1000, replace = TRUE)),
  n_var = rnorm(1000, mean = 0, sd = 1),
  resp = rnorm(1000, mean = 10, sd = 3))

# Run model
mod = brms_model(Response = "resp", 
                 FixedEffect = c("f_var","n_var"), 
                 RandomEffect = "group", 
                 Family = "gaussian", 
                 Data = md)


# Variance decomposition
var_decomp(mod)

}
