% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarDecomp-package.R
\docType{package}
\name{VarDecomp-package}
\alias{VarDecomp}
\alias{VarDecomp-package}
\title{VarDecomp: Variance Decomposition For Brms Models}
\description{
This package contains functions for variance decomposition of brms models, as well as functions for model fit checks, model comparisons and visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gabewinter/VarDecomp}
  \item Report bugs at \url{https://github.com/gabewinter/VarDecomp/issues}
}

}
\author{
\strong{Maintainer}: Gabe Winter \email{gabewinter@gmail.com} (\href{https://orcid.org/0000-0001-5984-9410}{ORCID})

}
\keyword{internal}
