% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varDecomp.R
\name{varDecomp}
\alias{varDecomp}
\title{Variance decomposition}
\usage{
varDecomp(data1, data2, formula, weight, ...)
}
\arguments{
\item{data1}{First dataset}

\item{data2}{Second dataset}

\item{formula}{Model formula}

\item{weight}{Weight variable}

\item{...}{Additional arguments for IPF.}
}
\value{
A list of decompositions
}
\description{
Variance decomposition
}
