% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_model.R
\name{brms_model}
\alias{brms_model}
\title{Run a brms model}
\usage{
brms_model(
  Data,
  Response,
  FixedEffect,
  RandomEffect = NULL,
  RandomSlope = NULL,
  Chainset = 1,
  Family = "gaussian",
  Seed = NULL,
  Trials = NA,
  PriorSamples = TRUE
)
}
\arguments{
\item{Data}{A data frame containing the data - covariates should be centered to a the mean or to a meaningful zero (see Schielzeth H. 2010. Simple means to improve the interpretability of regression coefficients. Methods Ecol Evol. 1:103–113. doi:10.1111/j.2041-210X.2010.00012.x.).}

\item{Response}{String with the name of the column in Data containing the response variable (e.g. "mass").}

\item{FixedEffect}{String with the name of the column in Data containing the fixed effect variable (e.g. "height"). To add multiple fixed effects, use c() (e.g. c("height", "sex")).}

\item{RandomEffect}{String with the name of the column in Data containing the random effect variable (e.g. "species"). The current package version allows the use of a single random effect.}

\item{RandomSlope}{String with the name of the column in Data containing the covariate to be added as a random slope (e.g. "height"). The current package version allows the use of a single random slope.}

\item{Chainset}{Defines the number of iterations. Start with Chainset = 1 and increase as needed until convergence. The value of Chainsetis multiplied by 15000 in warmup, 30000 in iterations and 15 in thin intervale. For quick tests use Chainset = 0 (warmup=10; iter=110; thin=10; chains=2)}

\item{Family}{String to define the family function in the brms model. Current supported families: "gaussian", "binomial", "poisson".}

\item{Seed}{Numeric and optional. Set a seed in order to repeat the results from the model when running it more than once.}

\item{Trials}{The total number of trials in a binomial model. The number of successes should be imputed on Response.}

\item{PriorSamples}{Logical value that defines if brmsfit will contain the priors used. Default is set to \code{FALSE}, which does not includ the priors in the brmsfit.}
}
\value{
Returns a brmsfit
}
\description{
Run a brms model
}
\examples{

md = dplyr::starwars

# Centering variables
md = md \%>\% 
  dplyr::select(mass, sex, height, species) \%>\% 
  dplyr::mutate(mass = log(mass),
         sex = dplyr::recode(sex, "male" = 1, 
                      "female" = -1, 
                      "hermaphroditic" = 0,
                      "none" = as.numeric(NA)))
  
  
mod = brms_model(Chainset = 0, 
           Response = "mass", 
           FixedEffect = c("sex","height"), 
           RandomEffect = "species", 
           RandomSlope = "height",
           Family = "gaussian", 
           Data = md, 
           Seed = 0405)

print(mod)

plot(mod)

}
