% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_slopes.R
\name{compare_slopes}
\alias{compare_slopes}
\title{Compare slope values of different brms models}
\usage{
compare_slopes(..., Slope)
}
\arguments{
\item{Slope}{A string containing the name of the covariate to have the slopes compared between models}

\item{brmsfit}{A list of brms models to be compared}
}
\value{
Returns a table of values with the log(slope model 1/slope model 2)
}
\description{
Compare slope values of different brms models
}
\examples{

md = dplyr::starwars

# Centering variables
md = md \%>\% 
  dplyr::select(mass, sex, height, species) \%>\% 
  dplyr::mutate(mass = log(mass),
         sex = dplyr::recode(sex, "male" = 1, 
                      "female" = -1, 
                      "hermaphroditic" = 0,
                      "none" = as.numeric(NA)))
  
  
mod1 = brms_model(Chainset = 0,
                  Response = "mass", 
           FixedEffect = "sex", 
           Family = "gaussian", 
           Data = md)

mod2 = brms_model(Chainset = 0,
                  Response = "mass", 
           FixedEffect = "sex",
           RandomEffect = "species",
           RandomSlope = "sex",
           Family = "gaussian", 
           Data = md)

compare_slopes(mod1, mod2, Slope = "sex")
}
