% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod.R
\docType{data}
\name{mod}
\alias{mod}
\title{Example model with fixed effects only}
\format{
A \code{brmsfit} object.
}
\source{
Generated using the \code{brms_model} function.
}
\usage{
mod
}
\description{
Example model with fixed effects only
}
\keyword{datasets}
