% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_examples-mod.R
\docType{data}
\name{mod}
\alias{mod}
\title{Model example that can be used to test other functions of \code{VarDecomp}. A model with fixed effects only.}
\format{
An object of class \code{brmsfit} with the following components:
\describe{
\item{formula}{The formula or formulas specifying the model.}
\item{data}{The data frame containing the variables used in the model.}
\item{prior}{The prior distributions for the model parameters.}
\item{sample_prior}{The prior samples, if \code{sample_prior} was set to \code{TRUE} during fitting.}
\item{stanmodel}{The Stan model code used for fitting.}
\item{stanfit}{The fitted model object returned by Stan.}
\item{algorithm}{The algorithm used for fitting the model, typically 'sampling' or 'meanfield'.}
\item{family}{The family object specifying the model family (e.g., Gaussian, Bernoulli).}
\item{data2}{Additional data used in the model, if any.}
\item{ranef}{A list of random effects structures in the model.}
\item{autocor}{The autocorrelation structure, if any, used in the model.}
\item{cov_ranef}{The covariance matrix of the random effects.}
\item{smooth}{A list of smooth terms in the model, if any.}
\item{stanvars}{A list of Stan variables defined in the model code.}
\item{inits}{Initial values used for the parameters in the Stan model.}
\item{fun}{The function used for sampling or variational inference.}
\item{stan_args}{A list of arguments passed to the Stan function.}
\item{control}{A list of control parameters for the Stan sampling algorithm.}
\item{mcmc_levs}{A list of MCMC settings used in the model fitting.}
\item{converged}{A logical indicating if the model has converged.}
\item{loo}{The leave-one-out cross-validation (LOO) results for the model, if computed.}
\item{waic}{The Watanabe-Akaike Information Criterion (WAIC) results for the model, if computed.}
\item{log_lik}{The log-likelihood values for the model, if computed.}
\item{fitted}{The fitted values from the model.}
\item{residuals}{The residuals from the model.}
\item{summary}{A summary of the model, including parameter estimates and diagnostics.}
\item{posterior_samples}{A matrix of posterior samples for the model parameters.}
\item{diagnostics}{Diagnostic information for the model, including convergence diagnostics.}
\item{user_data}{Any additional user data stored in the model object.}
}
}
\source{
Generated using the \code{brms_model} function.
}
\usage{
data(mod, package = "VarDecomp")
}
\description{
Model example that can be used to test other functions of \code{VarDecomp}. A model with fixed effects only.
}
\seealso{
\code{\link[brms]{brmsfit}} for more details on the \code{brmsfit} object.
}
\keyword{datasets}
