% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cov.R
\docType{data}
\name{mod_cov}
\alias{mod_cov}
\title{model with fixed effects and a random effect with a covariance matrix}
\format{
A \code{brmsfit} object.
}
\source{
Generated using the \code{brms_cov_model} function.
}
\usage{
mod_cov
}
\description{
model with fixed effects and a random effect with a covariance matrix
}
\keyword{datasets}
