% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_re.R, R/model_examples-mod_re.R,
%   R/model_examples.R
\docType{data}
\name{mod_RE}
\alias{mod_RE}
\title{Example model with fixed and random effects}
\format{
A \code{brmsfit} object.

A \code{brmsfit} object.

A \code{brmsfit} object.
}
\source{
Generated using the \code{brms_model} function.

Generated using the \code{brms_model} function.

Generated using the \code{brms_model} function.
}
\usage{
mod_RE

mod_RE

mod_RE
}
\description{
Example model with fixed and random effects

Model example that can be used to test other functions of \code{VarDecomp}. A model with fixed and random effects

Model example that can be used to test other functions of \code{VarDecomp}. A model with fixed and random effects
}
\keyword{datasets}
