% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{model_fit}
\alias{model_fit}
\title{Perform model fit checks for brms models}
\usage{
model_fit(brmsfit, Group = NULL, Prior = FALSE)
}
\arguments{
\item{brmsfit}{The output of a brms model. You can use VarDecomp::brms_model() to produce a brmsfit.}

\item{Group}{A string containing the name of a grouping variable for the visualization of a posterior predictive check plot (e.g. "sex").}

\item{Prior}{A logical argument defining whether the \code{brmsfit} contains prior samples. If set to \code{TRUE} it will produce plots comparing the log distributions of priors and posterior samples for each covariate.}
}
\value{
Returns a list containing (a) the maximum R-hat value, (b) the minimum effective sample size, (c) traceplots, (d) posterior predictive check plots, and (e) prior and posterior sample plots (if priors are available).
}
\description{
Perform model fit checks for brms models
}
\examples{
md = dplyr::starwars

# Centering variables
md = md \%>\% 
  dplyr::select(mass, sex, height, species) \%>\% 
  dplyr::mutate(mass = log(mass),
         sex = dplyr::recode(sex, "male" = 1, 
                      "female" = -1, 
                      "hermaphroditic" = 0,
                      "none" = as.numeric(NA)))


# Without random effects

mod = brms_model(Chainset= 2,
                 Response = "mass", 
                 FixedEffect = c("sex","height"), 
                 Family = "gaussian", 
                 Data = md,
                 PriorSamples = TRUE)

model_fit(mod, Group = "sex", Prior = TRUE)
}
