% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intervals.R
\name{plot_intervals}
\alias{plot_intervals}
\title{Plot probability density intervals for fixed effects of brms models}
\usage{
plot_intervals(brmsfit)
}
\arguments{
\item{brmsfit}{The output of a brms model. You can use VarDecomp::brms_model() to produce a brmsfit.}
}
\value{
Returns a density plot with z-transformed fixed effect estimates.
}
\description{
Plot probability density intervals for fixed effects of brms models
}
\examples{
\dontrun{
md = dplyr::starwars

# Centering variables
md = md \%>\% 
  dplyr::select(mass, sex, species) \%>\% 
  dplyr::mutate(mass = log(mass),
         sex = dplyr::recode(sex, "male" = 1, 
                      "female" = -1, 
                      "hermaphroditic" = 0,
                      "none" = as.numeric(NA)))
  
  
mod = brms_model(Chainset = 2,
                 Response = "mass", 
                 FixedEffect = "sex", 
                 RandomEffect = "species",
                 Family = "gaussian", 
                 Data = md)

plot_intervals(mod)

}

}
