% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_decomp.R
\name{var_decomp}
\alias{var_decomp}
\title{Variance decomposition for brms models}
\usage{
var_decomp(brmsfit)
}
\arguments{
\item{brmsfit}{The output of a brms model. You can use VarDecomp::brms_model() to produce a brmsfit.}
}
\value{
Returns a data frame with the summaries of posterior estimates.
}
\description{
Variance decomposition for brms models
}
\examples{
#var_decomp(mod_RS)
}
